combine methodgen and typegen in generator.py

+++ src/methods/add_sticker_to_set.rs
@@ -1,7 +1,8 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct addStickerToSet{
-    pub user_id: i64,
+    pub user_id: i32,
     pub name: String,
     pub png_sticker: Option<InputFile>,
     pub tgs_sticker: Option<InputFile>,
