+++ src/methods/answer_inline_query.rs
@@ -1,9 +1,10 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct answerInlineQuery{
     pub inline_query_id: String,
     pub results: Vec<InlineQueryResult>,
-    pub cache_time: Option<i64>,
+    pub cache_time: Option<i32>,
     pub is_personal: Option<bool>,
     pub next_offset: Option<String>,
     pub switch_pm_text: Option<String>,
