+++ src/methods/create_new_sticker_set.rs
@@ -1,7 +1,8 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct createNewStickerSet{
-    pub user_id: i64,
+    pub user_id: i32,
     pub name: String,
     pub title: String,
     pub png_sticker: Option<InputFile>,
