+++ src/methods/delete_chat_sticker_set.rs
@@ -1,5 +1,6 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct deleteChatStickerSet{
-    pub chat_id: i64
+    pub chat_id: i32
 }
