+++ src/methods/delete_message.rs
@@ -1,6 +1,7 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct deleteMessage{
-    pub chat_id: i64,
-    pub message_id: i64
+    pub chat_id: i32,
+    pub message_id: i32
 }
