+++ src/methods/edit_message_live_location.rs
@@ -1,13 +1,14 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct editMessageLiveLocation{
-    pub chat_id: Option<i64>,
-    pub message_id: Option<i64>,
+    pub chat_id: Option<i32>,
+    pub message_id: Option<i32>,
     pub inline_message_id: Option<String>,
     pub latitude: f64,
     pub longitude: f64,
     pub horizontal_accuracy: Option<f64>,
-    pub heading: Option<i64>,
-    pub proximity_alert_radius: Option<i64>,
+    pub heading: Option<i32>,
+    pub proximity_alert_radius: Option<i32>,
     pub reply_markup: Option<InlineKeyboardMarkup>
 }
