+++ src/methods/forward_message.rs
@@ -1,8 +1,9 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct forwardMessage{
-    pub chat_id: i64,
-    pub from_chat_id: i64,
+    pub chat_id: i32,
+    pub from_chat_id: i32,
     pub disable_notification: Option<bool>,
-    pub message_id: i64
+    pub message_id: i32
 }
