+++ src/methods/get_chat_member.rs
@@ -1,6 +1,7 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct getChatMember{
-    pub chat_id: i64,
-    pub user_id: i64
+    pub chat_id: i32,
+    pub user_id: i32
 }
