+++ src/methods/get_chat_members_count.rs
@@ -1,5 +1,6 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct getChatMembersCount{
-    pub chat_id: i64
+    pub chat_id: i32
 }
