+++ src/methods/get_game_high_scores.rs
@@ -1,8 +1,9 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct getGameHighScores{
-    pub user_id: i64,
-    pub chat_id: Option<i64>,
-    pub message_id: Option<i64>,
+    pub user_id: i32,
+    pub chat_id: Option<i32>,
+    pub message_id: Option<i32>,
     pub inline_message_id: Option<String>
 }
