+++ src/methods/get_updates.rs
@@ -1,8 +1,9 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct getUpdates{
-    pub offset: Option<i64>,
-    pub limit: Option<i64>,
-    pub timeout: Option<i64>,
+    pub offset: Option<i32>,
+    pub limit: Option<i32>,
+    pub timeout: Option<i32>,
     pub allowed_updates: Option<Vec<String>>
 }
