+++ src/methods/get_user_profile_photos.rs
@@ -1,7 +1,8 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct getUserProfilePhotos{
-    pub user_id: i64,
-    pub offset: Option<i64>,
-    pub limit: Option<i64>
+    pub user_id: i32,
+    pub offset: Option<i32>,
+    pub limit: Option<i32>
 }
