+++ src/methods/kick_chat_member.rs
@@ -1,7 +1,8 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct kickChatMember{
-    pub chat_id: i64,
-    pub user_id: i64,
-    pub until_date: Option<i64>
+    pub chat_id: i32,
+    pub user_id: i32,
+    pub until_date: Option<i32>
 }
