+++ src/methods/promote_chat_member.rs
@@ -1,8 +1,9 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct promoteChatMember{
-    pub chat_id: i64,
-    pub user_id: i64,
+    pub chat_id: i32,
+    pub user_id: i32,
     pub is_anonymous: Option<bool>,
     pub can_change_info: Option<bool>,
     pub can_post_messages: Option<bool>,
