+++ src/methods/restrict_chat_member.rs
@@ -1,8 +1,9 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct restrictChatMember{
-    pub chat_id: i64,
-    pub user_id: i64,
+    pub chat_id: i32,
+    pub user_id: i32,
     pub permissions: ChatPermissions,
-    pub until_date: Option<i64>
+    pub until_date: Option<i32>
 }
