+++ src/methods/send_chat_action.rs
@@ -1,6 +1,7 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct sendChatAction{
-    pub chat_id: i64,
+    pub chat_id: i32,
     pub action: String
 }
