+++ src/methods/send_document.rs
@@ -1,7 +1,8 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct sendDocument{
-    pub chat_id: i64,
+    pub chat_id: i32,
     pub document: InputFile,
     pub thumb: Option<InputFile>,
     pub caption: Option<String>,
@@ -9,7 +10,7 @@ pub struct sendDocument{
     pub caption_entities: Option<Vec<MessageEntity>>,
     pub disable_content_type_detection: Option<bool>,
     pub disable_notification: Option<bool>,
-    pub reply_to_message_id: Option<i64>,
+    pub reply_to_message_id: Option<i32>,
     pub allow_sending_without_reply: Option<bool>,
     pub reply_markup: Option<InlineKeyboardMarkup>
 }
