+++ src/methods/send_invoice.rs
@@ -1,7 +1,8 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct sendInvoice{
-    pub chat_id: i64,
+    pub chat_id: i32,
     pub title: String,
     pub description: String,
     pub payload: String,
@@ -11,9 +12,9 @@ pub struct sendInvoice{
     pub prices: Vec<LabeledPrice>,
     pub provider_data: Option<String>,
     pub photo_url: Option<String>,
-    pub photo_size: Option<i64>,
-    pub photo_width: Option<i64>,
-    pub photo_height: Option<i64>,
+    pub photo_size: Option<i32>,
+    pub photo_width: Option<i32>,
+    pub photo_height: Option<i32>,
     pub need_name: Option<bool>,
     pub need_phone_number: Option<bool>,
     pub need_email: Option<bool>,
@@ -22,7 +23,7 @@ pub struct sendInvoice{
     pub send_email_to_provider: Option<bool>,
     pub is_flexible: Option<bool>,
     pub disable_notification: Option<bool>,
-    pub reply_to_message_id: Option<i64>,
+    pub reply_to_message_id: Option<i32>,
     pub allow_sending_without_reply: Option<bool>,
     pub reply_markup: Option<InlineKeyboardMarkup>
 }
