+++ src/methods/send_location.rs
@@ -1,15 +1,16 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct sendLocation{
-    pub chat_id: i64,
+    pub chat_id: i32,
     pub latitude: f64,
     pub longitude: f64,
     pub horizontal_accuracy: Option<f64>,
-    pub live_period: Option<i64>,
-    pub heading: Option<i64>,
-    pub proximity_alert_radius: Option<i64>,
+    pub live_period: Option<i32>,
+    pub heading: Option<i32>,
+    pub proximity_alert_radius: Option<i32>,
     pub disable_notification: Option<bool>,
-    pub reply_to_message_id: Option<i64>,
+    pub reply_to_message_id: Option<i32>,
     pub allow_sending_without_reply: Option<bool>,
     pub reply_markup: Option<InlineKeyboardMarkup>
 }
