+++ src/methods/send_photo.rs
@@ -1,13 +1,14 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct sendPhoto{
-    pub chat_id: i64,
+    pub chat_id: i32,
     pub photo: InputFile,
     pub caption: Option<String>,
     pub parse_mode: Option<String>,
     pub caption_entities: Option<Vec<MessageEntity>>,
     pub disable_notification: Option<bool>,
-    pub reply_to_message_id: Option<i64>,
+    pub reply_to_message_id: Option<i32>,
     pub allow_sending_without_reply: Option<bool>,
     pub reply_markup: Option<InlineKeyboardMarkup>
 }
