+++ src/methods/send_venue.rs
@@ -1,7 +1,8 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct sendVenue{
-    pub chat_id: i64,
+    pub chat_id: i32,
     pub latitude: f64,
     pub longitude: f64,
     pub title: String,
@@ -11,7 +12,7 @@ pub struct sendVenue{
     pub google_place_id: Option<String>,
     pub google_place_type: Option<String>,
     pub disable_notification: Option<bool>,
-    pub reply_to_message_id: Option<i64>,
+    pub reply_to_message_id: Option<i32>,
     pub allow_sending_without_reply: Option<bool>,
     pub reply_markup: Option<InlineKeyboardMarkup>
 }
