+++ src/methods/set_chat_administrator_custom_title.rs
@@ -1,7 +1,8 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct setChatAdministratorCustomTitle{
-    pub chat_id: i64,
-    pub user_id: i64,
+    pub chat_id: i32,
+    pub user_id: i32,
     pub custom_title: String
 }
