+++ src/methods/set_chat_description.rs
@@ -1,6 +1,7 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct setChatDescription{
-    pub chat_id: i64,
+    pub chat_id: i32,
     pub description: Option<String>
 }
