+++ src/methods/set_chat_sticker_set.rs
@@ -1,6 +1,7 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct setChatStickerSet{
-    pub chat_id: i64,
+    pub chat_id: i32,
     pub sticker_set_name: String
 }
