+++ src/methods/set_chat_title.rs
@@ -1,6 +1,7 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct setChatTitle{
-    pub chat_id: i64,
+    pub chat_id: i32,
     pub title: String
 }
