+++ src/methods/set_sticker_position_in_set.rs
@@ -1,6 +1,7 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct setStickerPositionInSet{
     pub sticker: String,
-    pub position: i64
+    pub position: i32
 }
