+++ src/methods/set_sticker_set_thumb.rs
@@ -1,7 +1,8 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct setStickerSetThumb{
     pub name: String,
-    pub user_id: i64,
+    pub user_id: i32,
     pub thumb: Option<InputFile>
 }
