+++ src/methods/stop_poll.rs
@@ -1,7 +1,8 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct stopPoll{
-    pub chat_id: i64,
-    pub message_id: i64,
+    pub chat_id: i32,
+    pub message_id: i32,
     pub reply_markup: Option<InlineKeyboardMarkup>
 }
