+++ src/methods/unban_chat_member.rs
@@ -1,7 +1,8 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct unbanChatMember{
-    pub chat_id: i64,
-    pub user_id: i64,
+    pub chat_id: i32,
+    pub user_id: i32,
     pub only_if_banned: Option<bool>
 }
