+++ src/methods/unpin_chat_message.rs
@@ -1,6 +1,7 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct unpinChatMessage{
-    pub chat_id: i64,
-    pub message_id: Option<i64>
+    pub chat_id: i32,
+    pub message_id: Option<i32>
 }
