+++ src/methods/upload_sticker_file.rs
@@ -1,6 +1,7 @@
 use serde_json;
+
 #[derive(Debug, Serialize)]
 pub struct uploadStickerFile{
-    pub user_id: i64,
+    pub user_id: i32,
     pub png_sticker: InputFile
 }
