+++ src/types/audio.rs
@@ -4,11 +4,11 @@ use crate::types::{photo_size::PhotoSize}
 pub struct Audio{
     pub file_id: String,
     pub file_unique_id: String,
-    pub duration: i64,
+    pub duration: i32,
     pub performer: Option<String>,
     pub title: Option<String>,
     pub file_name: Option<String>,
     pub mime_type: Option<String>,
-    pub file_size: Option<i64>,
+    pub file_size: Option<i32>,
     pub thumb: Option<PhotoSize>
 }
