+++ src/types/chat_member.rs
@@ -21,5 +21,5 @@ pub struct ChatMember{
     pub can_send_polls: Option<bool>,
     pub can_send_other_messages: Option<bool>,
     pub can_add_web_page_previews: Option<bool>,
-    pub until_date: Option<i64>
+    pub until_date: Option<i32>
 }
