+++ src/types/contact.rs
@@ -3,6 +3,6 @@ pub struct Contact{
     pub phone_number: String,
     pub first_name: String,
     pub last_name: Option<String>,
-    pub user_id: Option<i64>,
+    pub user_id: Option<i32>,
     pub vcard: Option<String>
 }
