+++ src/types/document.rs
@@ -7,5 +7,5 @@ pub struct Document{
     pub thumb: Option<PhotoSize>,
     pub file_name: Option<String>,
     pub mime_type: Option<String>,
-    pub file_size: Option<i64>
+    pub file_size: Option<i32>
 }
