+++ src/types/file.rs
@@ -2,6 +2,6 @@
 pub struct File{
     pub file_id: String,
     pub file_unique_id: String,
-    pub file_size: Option<i64>,
+    pub file_size: Option<i32>,
     pub file_path: Option<String>
 }
