+++ src/types/game_high_score.rs
@@ -2,7 +2,7 @@ use crate::types::{user::User}
 
 #[derive(Debug, Deserialize)]
 pub struct GameHighScore{
-    pub position: i64,
+    pub position: i32,
     pub user: User,
-    pub score: i64
+    pub score: i32
 }
