+++ src/types/labeled_price.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Deserialize)]
 pub struct LabeledPrice{
     pub label: String,
-    pub amount: i64
+    pub amount: i32
 }
