+++ src/types/location.rs
@@ -3,7 +3,7 @@ pub struct Location{
     pub longitude: f64,
     pub latitude: f64,
     pub horizontal_accuracy: Option<f64>,
-    pub live_period: Option<i64>,
-    pub heading: Option<i64>,
-    pub proximity_alert_radius: Option<i64>
+    pub live_period: Option<i32>,
+    pub heading: Option<i32>,
+    pub proximity_alert_radius: Option<i32>
 }
