+++ src/types/message_entity.rs
@@ -3,8 +3,8 @@ use crate::types::{user::User}
 #[derive(Debug, Deserialize)]
 pub struct MessageEntity{
     pub type: String,
-    pub offset: i64,
-    pub length: i64,
+    pub offset: i32,
+    pub length: i32,
     pub url: Option<String>,
     pub user: Option<User>,
     pub language: Option<String>
