+++ src/types/passport_file.rs
@@ -2,6 +2,6 @@
 pub struct PassportFile{
     pub file_id: String,
     pub file_unique_id: String,
-    pub file_size: i64,
-    pub file_date: i64
+    pub file_size: i32,
+    pub file_date: i32
 }
