+++ src/types/photo_size.rs
@@ -2,7 +2,7 @@
 pub struct PhotoSize{
     pub file_id: String,
     pub file_unique_id: String,
-    pub width: i64,
-    pub height: i64,
-    pub file_size: Option<i64>
+    pub width: i32,
+    pub height: i32,
+    pub file_size: Option<i32>
 }
