+++ src/types/poll_option.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Deserialize)]
 pub struct PollOption{
     pub text: String,
-    pub voter_count: i64
+    pub voter_count: i32
 }
