+++ src/types/response_parameters.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Deserialize)]
 pub struct ResponseParameters{
     pub migrate_to_chat_id: Option<i64>,
-    pub retry_after: Option<i64>
+    pub retry_after: Option<i32>
 }
