+++ src/types/user_profile_photos.rs
@@ -2,6 +2,6 @@ use crate::types::{photo_size::PhotoSize}
 
 #[derive(Debug, Deserialize)]
 pub struct UserProfilePhotos{
-    pub total_count: i64,
+    pub total_count: i32,
     pub photos: Vec<Array> of PhotoSize
 }
