+++ src/types/video.rs
@@ -4,11 +4,11 @@ use crate::types::{photo_size::PhotoSize}
 pub struct Video{
     pub file_id: String,
     pub file_unique_id: String,
-    pub width: i64,
-    pub height: i64,
-    pub duration: i64,
+    pub width: i32,
+    pub height: i32,
+    pub duration: i32,
     pub thumb: Option<PhotoSize>,
     pub file_name: Option<String>,
     pub mime_type: Option<String>,
-    pub file_size: Option<i64>
+    pub file_size: Option<i32>
 }
