+++ src/types/video_note.rs
@@ -4,8 +4,8 @@ use crate::types::{photo_size::PhotoSize}
 pub struct VideoNote{
     pub file_id: String,
     pub file_unique_id: String,
-    pub length: i64,
-    pub duration: i64,
+    pub length: i32,
+    pub duration: i32,
     pub thumb: Option<PhotoSize>,
-    pub file_size: Option<i64>
+    pub file_size: Option<i32>
 }
