+++ src/types/voice.rs
@@ -2,7 +2,7 @@
 pub struct Voice{
     pub file_id: String,
     pub file_unique_id: String,
-    pub duration: i64,
+    pub duration: i32,
     pub mime_type: Option<String>,
-    pub file_size: Option<i64>
+    pub file_size: Option<i32>
 }
