+++ src/types/webhook_info.rs
@@ -2,10 +2,10 @@
 pub struct WebhookInfo{
     pub url: String,
     pub has_custom_certificate: bool,
-    pub pending_update_count: i64,
+    pub pending_update_count: i32,
     pub ip_address: Option<String>,
-    pub last_error_date: Option<i64>,
+    pub last_error_date: Option<i32>,
     pub last_error_message: Option<String>,
-    pub max_connections: Option<i64>,
+    pub max_connections: Option<i32>,
     pub allowed_updates: Option<Vec<String>>
 }
